Restaurar y renombrar componente WelcomeMessage.jsx
