+++ frontend/src/components/WelcomeMessage.jsx
@@ -1,11 +1,10 @@
 import { motion } from "framer-motion";
 
-
 function capitalizeFirstLetter(name) {
   if (!name) return "";
-  return name.charAt(0).toUpperCase() + name.slice(1);
+  const s = String(name).trim();
+  return s.charAt(0).toUpperCase() + s.slice(1);
 }
-
 
 function getDayEmoji() {
   const day = new Date().getDay(); // 0 = domingo, 1 = lunes, etc.
@@ -13,7 +12,7 @@ function getDayEmoji() {
   return emojis[day] || "😊";
 }
 
-export function WelcomeMessage({ name }) {
+export function WelcomeMessage({ name = "" }) {
   const displayName = capitalizeFirstLetter(name);
   const emoji = getDayEmoji();
 
@@ -24,36 +23,40 @@ export function WelcomeMessage({ name }) {
       initial={{ opacity: 0, y: -20 }}
       animate={{ opacity: 1, y: 0 }}
       transition={{ duration: 0.6, type: "spring" }}
-      className="mt-6 text-center bg-[#2A1F16] py-10 rounded-xl shadow-lg"
+      className="bg-[#1a1a1a] min-h-screen flex items-center justify-center text-center"
     >
-      <h3
-        className="text-5xl font-extrabold text-[#FFF200] drop-shadow-[0_0_8px_#FFF200]"
-        style={{
-          textShadow:
-            "0 0 5px #FFF200, 0 0 10px #E8BB78, 0 0 20px #E8BB78, 0 0 30px #E8BB78",
-        }}
-      >
-        ¡Bienvenido,{" "}
-        <motion.span
-          aria-label={`Nombre del usuario: ${displayName || "Usuario"}`}
-          initial={{ scale: 0.8 }}
-          animate={{
-            scale: 1,
+      <div>
+        {/* Título principal */}
+        <h3
+          className="text-6xl font-extrabold text-[#FFF200]"
+          style={{
             textShadow:
-              "0 0 10px #E8BB78, 0 0 20px #FFF200, 0 0 30px #FFF200",
+              "2px 2px 4px #987A64, 0 0 15px #E8BB78, 0 0 30px #FFF200",
           }}
-          transition={{ type: "spring", stiffness: 300, duration: 1 }}
-          className="text-[#E8BB78]"
         >
-          {displayName || "Usuario"}
-        </motion.span>{" "}
-        <span aria-hidden="true">{emoji}</span>
-      </h3>
+          ¡Bienvenido,{" "}
+          <motion.span
+            aria-label={`Nombre del usuario: ${displayName || "Usuario"}`}
+            initial={{ scale: 0.9 }}
+            animate={{
+              scale: 1,
+              textShadow:
+                "2px 2px 4px #987A64, 0 0 20px #E8BB78, 0 0 40px #FFF200",
+            }}
+            transition={{ type: "spring", stiffness: 300, duration: 1 }}
+            className="text-[#E8BB78]"
+          >
+            {displayName || "Usuario"}
+          </motion.span>{" "}
+          <span aria-hidden="true">{emoji}</span>
+        </h3>
 
-      <p className="text-[#987A64] font-medium mt-4 text-xl">
-        Nos alegra verte de nuevo en{" "}
-        <span className="font-bold text-[#E8BB78]">Recselpermadispro</span>
-      </p>
+        {/* Subtítulo */}
+        <p className="mt-4 text-lg font-medium text-[#987A64]">
+          Nos alegra verte de nuevo en{" "}
+          <span className="font-bold text-[#E8BB78]">Recselpermadispro</span>
+        </p>
+      </div>
     </motion.div>
   );
 }
